+++ subscriptiongen/src/prototypes/connection.rs
@@ -29,7 +29,7 @@ impl ConnectionData {
         quote! { connection: TrackedConnection, }
     }
 
-    pub fn gen_update(bitmask: usize) -> proc_macro2::TokenStream {
+    pub fn gen_update(bitmask: u128) -> proc_macro2::TokenStream {
         let bitmask_lit = syn::LitInt::new(&bitmask.to_string(), Span::call_site());
         quote! { 
             if self.match_data.matching_by_bitmask(#bitmask_lit) {
